+++ pyburgers/physics/sgs/sgs_smagcon.py
@@ -50,7 +50,7 @@ class SmagConstant(SGS):
         """
         super().__init__(input_obj, spectral)
         self.logger: logging.Logger = get_logger("SGS")
-        self.logger.info("--- Using the Smagorinsky model")
+        self.logger.info("--- using the Smagorinsky model")
 
     def compute(
         self, u: np.ndarray, dudx: np.ndarray, tke_sgs: np.ndarray | float, dt: float